--- packages/promo-pages/next.config.ts
@@ -1,7 +1,0 @@
-import type {NextConfig} from 'next';
-
-const nextConfig: NextConfig = {
-	output: 'export',
-};
-
-export default nextConfig;
