+++ packages/promo-pages/package.json
@@ -1,12 +1,10 @@
 {
 	"name": "@remotion/promo-pages",
-	"version": "0.1.0",
+	"version": "4.0.248",
 	"private": true,
 	"scripts": {
-		"dev": "next dev",
-		"build": "next build",
-		"start": "next start",
-		"lint": "next lint"
+		"make": "bun bundle.ts",
+		"lint": "tsc"
 	},
 	"dependencies": {
 		"@remotion/animated-emoji": "workspace:*",
@@ -14,14 +12,15 @@
 		"@remotion/lottie": "workspace:*",
 		"@remotion/paths": "workspace:*",
 		"@remotion/player": "workspace:*",
+		"@remotion/eslint-config-internal": "workspace:*",
 		"@remotion/shapes": "workspace:*",
 		"create-video": "workspace:*",
 		"hls.js": "^0.14.17",
-		"next": "15.1.4",
 		"polished": "4.3.1",
 		"react": "19.0.0",
 		"react-dom": "19.0.0",
 		"remotion": "workspace:*",
+		"vite": "^5.4.6",
 		"zod": "3.22.3"
 	},
 	"devDependencies": {
@@ -33,7 +32,8 @@
 		"eslint-config-next": "15.0.4",
 		"plyr": "^3.7.3",
 		"postcss": "8.4.47",
-		"tailwindcss": "3.4.13",
+		"tailwindcss": "4.0.0-beta.9",
+		"@tailwindcss/cli": "4.0.0-beta.9",
 		"tailwindcss-scoped-preflight": "3.4.10"
 	}
 }
