--- packages/promo-pages/src/app/globals.css
@@ -1,41 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-	font-family: Arial, Helvetica, sans-serif;
-}
-
-@font-face {
-	font-family: 'GTPlanar';
-	src: url('/gt-planar-bold.woff2') format('woff2');
-	font-weight: 700;
-	font-style: normal;
-	font-feature-settings: 'ss03' on;
-}
-@font-face {
-	font-family: 'GTPlanar';
-	src: url('/gt-planar-medium.woff2') format('woff2');
-	font-weight: 500;
-	font-style: normal;
-	font-feature-settings: 'ss03' on;
-}
-@font-face {
-	font-family: 'GTPlanar';
-	src: url('/gt-planar-regular.woff2') format('woff2');
-	font-weight: 400;
-	font-style: normal;
-	font-feature-settings: 'ss03' on;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-	letter-spacing: -0.025em;
-	word-spacing: -0.025em;
-	font-family: 'GTPlanar';
-	font-feature-settings: 'ss03' on;
-}
