--- packages/promo-pages/src/app/page.tsx
@@ -1,5 +1,0 @@
-import NewLanding from '@/components/homepage/Homepage';
-
-export default function Home() {
-	return <NewLanding />;
-}
